@@ -1,11 +1,11 @@
-import React from 'react';
+import PropTypes from 'prop-types';
 
 function App(props) {
   return (props.children);
 }
 
 App.propTypes = {
-  children: React.PropTypes.element.isRequired
+  children: PropTypes.element.isRequired,
 };
 
 export default App;
