@@ -33,7 +33,7 @@ const adapter = {
     }
 
     return fn(null, null);
-  }
+  },
 };
 
 export default adapter;
